Moves AboutPreferencesFragment out of AboutPreferencesActivity
